fix: refactor entities 2.0

+++ src/main/java/org/cheeseapp/repos/ClientRepo.java
@@ -1,8 +1,7 @@
 package org.cheeseapp.repos;
 
-
 import org.cheeseapp.domain.Client;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClientRepo extends CrudRepository<Client,Integer> {
+public interface ClientRepo extends CrudRepository<Client, Long> {
 }
